requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.1
SQLAlchemy==2.0.44
psycopg[binary]==3.1.18
gunicorn==21.2.0